requirements.txt
streamlit
torch
torchvision
pillow
gdown
matplotlib
numpy
